Added bonus part to message-chain exercise, and combined the solutions into one file
