+++ mpi/hello-world/README.md
@@ -11,8 +11,12 @@
 
 3. (Bonus) Sometimes it is convenient to find out in which nodes the different MPI
    processes are running. MPI has a function `MPI_Get_processor_name` for quering the
-   name of the node. Use the function in your program (you can search for the interface
-   or check the MPI reference in the course material or man pages `man MPI_Get_processor_name`)
-   and print out the node name in all processes.
-   You can request specific number of nodes with the `--nodes` Slurm option,
-   investigate how the results look with different values.
+   name of the node. Use this function in your program and print out the node name in all processes.
+   Use any of the following sources to find the usage of `MPI_Get_processor_name`:
+
+   - The official MPI specification: https://www.mpi-forum.org/docs/
+   - Manual pages with Linux `man` command: `man MPI_Get_processor_name`
+   - Search engine of your choice. These usually bring you to the docs of OpenMPI or MPICH, both excellent resources.
+
+   When submitting your program to Slurm, you can request a specific number of nodes with the `--nodes` Slurm option.
+   Investigate how the results look with different values.
