--- mpi/message-chain/solution/chain-send-recv.cpp
@@ -1,91 +1,0 @@
-#include <cstdio>
-#include <vector>
-#include <mpi.h>
-
-void print_ordered(double t);
-
-int main(int argc, char *argv[])
-{
-    constexpr int numElements = 10000000;
-    std::vector<int> message(numElements);
-    std::vector<int> receiveBuffer(numElements);
-    MPI_Status status;
-
-    MPI_Init(&argc, &argv);
-
-    int rank, ntasks;
-    MPI_Comm_size(MPI_COMM_WORLD, &ntasks);
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-
-    // Initialize buffers
-    for (int i = 0; i < numElements; i++)
-    {
-        message[i] = rank;
-        receiveBuffer[i] = -1;
-    }
-
-    // Set source and destination ranks, handling boundaries with MPI_PROC_NULL
-    int source = rank - 1;
-    int destination = rank + 1;
-
-    // First rank receives from no one
-    if (rank == 0)
-    {
-        source = MPI_PROC_NULL;
-    }
-    // Last rank sends to no one
-    if (rank >= ntasks - 1)
-    {
-        destination = MPI_PROC_NULL;
-    }
-
-
-    // Start measuring the time spent in communication
-    MPI_Barrier(MPI_COMM_WORLD);
-    double t0 = MPI_Wtime();
-
-    // Send messages. Note that the send tag must be valid (>= 0) even if our destination is MPI_PROC_NULL
-    int sendTag = rank + 1;
-    MPI_Send(message.data(), numElements, MPI_INT, destination, sendTag, MPI_COMM_WORLD);
-
-    printf("Sender: %d. Sent elements: %d. Tag: %d. Receiver: %d\n",
-           rank, numElements, sendTag, destination);
-
-    // Receive messages
-    int recvTag = rank;
-    MPI_Recv(receiveBuffer.data(), numElements, MPI_INT, source, recvTag, MPI_COMM_WORLD, &status);
-
-    printf("Receiver: %d. first element %d.\n",
-           rank, receiveBuffer[0]);
-
-    // Finalize measuring the time and print it out
-    double t1 = MPI_Wtime();
-    MPI_Barrier(MPI_COMM_WORLD);
-    fflush(stdout);
-
-    print_ordered(t1 - t0);
-
-    MPI_Finalize();
-    return 0;
-}
-
-void print_ordered(double t)
-{
-    int i, rank, ntasks;
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    MPI_Comm_size(MPI_COMM_WORLD, &ntasks);
-
-    if (rank == 0)
-    {
-        printf("Time elapsed in rank %2d: %6.3f\n", rank, t);
-        for (i = 1; i < ntasks; i++)
-        {
-            MPI_Recv(&t, 1, MPI_DOUBLE, i, 11, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
-            printf("Time elapsed in rank %2d: %6.3f\n", i, t);
-        }
-    }
-    else
-    {
-        MPI_Send(&t, 1, MPI_DOUBLE, 0, 11, MPI_COMM_WORLD);
-    }
-}
